+++ geographic_services/service_area/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ServiceAreaConfig(AppConfig):
+
+    name = 'geographic_services.service_area'