+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Geographic Services Changelog
 
-## [NEXT_RELEASE] ## [0.2.0] - 2022-12-04 13:34
+## [NEXT_RELEASE]
+
+### Added
+
+   - Service area app module
+   - Endpoint to get service areas based on latitude and longitude
+   - Cache configuration and decorators
+
+## [0.2.0] - 2022-12-04 13:34
 
 ### Changed